refactor: api 에러 디버깅 추가
