+++ src/app/(order-flow)/cart/page.tsx
@@ -1,5 +1,7 @@
 import { Cart } from '@/features/cart';
 
-export default function CartPage() {
+export const dynamic = 'force-dynamic';
+
+export default async function CartPage() {
   return <Cart />;
 }
