+++ src/app/(order-flow)/checkout/[id]/page.tsx
@@ -1,14 +1,12 @@
 import { Checkout } from '@/features/order';
-import { fetchOrderSheetById } from '@/features/order-sheet/apis';
-import { redirect } from 'next/navigation';
+import { fetchOrderSheetById } from '@/features/order-sheet';
 
 export default async function CheckoutPage({
   params,
 }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
-  const orderSheet = await fetchOrderSheetById(id);
 
-  if (!orderSheet.orderSheetId) redirect('/cart');
+  const orderSheet = await fetchOrderSheetById(id);
 
   return <Checkout orderSheet={orderSheet} />;
 }
