+++ src/app/api/orders/[orderId]/route.ts
@@ -37,16 +37,13 @@ export async function GET(
     .eq('address_id', orders[0].address_id);
 
   if (addressesError) {
-    console.log('addressesError', addressesError);
+    console.error('addressesError', addressesError);
     return NextResponse.json({ error: '주문 주소 조회 실패' }, { status: 500 });
   }
 
   const response = {
     orderId,
-    orderDate: format(orders[0].order_date, 'YYYY-MM-DD HH:mm:ss', {
-      locale: ko,
-    }),
-
+    orderDate: orders[0].order_date,
     payment: {
       paymentKey: payment[0].payment_key,
       paymentMethod: payment[0].payment_method,
